"Model" first design approach implemented. Entering experimental stage: the following needs fully testing/fixing/completing/improving: 1. The Model API (please refer to my Model.java inline comments). 2. Improving the examples found in examples.Model (just observe them, guess and be creative ;) ).
On the implementation side of notes: What was then required for the
Scene to support the implementation of Models? a. A
scene.pickingBuffer() (the idea found online), b. automatically calling
into the pickingBuffer() from within P5.post(), c. implementing some
cutomized (picking) MatrixHelpers (to sync the main scene.pg() and
pickingBuffer() eye-frames, without the need for a extra-scene object),
and d. implementing two static scene methods to be able to apply a frame
transformation into a given PGraphics (see
scene.applyWorldTransformation(PGraphics pg)). Just as easy and simple
for one to love it ;)

I'm optimistic madels have to see the light.